+++ testlib/src/main/java/com/diffplug/spotless/StepHarnessWithFile.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2023 DiffPlug
+ * Copyright 2016-2024 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,13 +26,12 @@ import org.assertj.core.api.AbstractStringAssert;
 import org.assertj.core.api.Assertions;
 
 /** An api for testing a {@code FormatterStep} that depends on the File path. */
-public class StepHarnessWithFile implements AutoCloseable {
-	private final Formatter formatter;
+public class StepHarnessWithFile extends StepHarnessBase<StepHarnessWithFile> {
 	private final ResourceHarness harness;
 
 	private StepHarnessWithFile(ResourceHarness harness, Formatter formatter) {
+		super(formatter);
 		this.harness = Objects.requireNonNull(harness);
-		this.formatter = Objects.requireNonNull(formatter);
 	}
 
 	/** Creates a harness for testing steps which do depend on the file. */
@@ -53,17 +52,23 @@ public class StepHarnessWithFile implements AutoCloseable {
 	}
 
 	/** Asserts that the given element is transformed as expected, and that the result is idempotent. */
-	public StepHarnessWithFile test(File file, String before, String after) {
-		String actual = formatter.compute(LineEnding.toUnix(before), file);
+	public StepHarnessWithFile test(String filename, String before, String after) {
+		File file = harness.setFile(filename).toContent(before);
+		String actual = formatter().compute(LineEnding.toUnix(before), file);
 		assertEquals(after, actual, "Step application failed");
 		return testUnaffected(file, after);
 	}
 
 	/** Asserts that the given element is idempotent w.r.t the step under test. */
 	public StepHarnessWithFile testUnaffected(File file, String idempotentElement) {
-		String actual = formatter.compute(LineEnding.toUnix(idempotentElement), file);
+		String actual = formatter().compute(LineEnding.toUnix(idempotentElement), file);
 		assertEquals(idempotentElement, actual, "Step is not idempotent");
 		return this;
+	}
+
+	/** Asserts that the given element is idempotent w.r.t the step under test. */
+	public StepHarnessWithFile testUnaffected(String file, String idempotentElement) {
+		return testUnaffected(harness.setFile(file).toContent(idempotentElement), idempotentElement);
 	}
 
 	/** Asserts that the given elements in  the resources directory are transformed as expected. */
@@ -73,8 +78,7 @@ public class StepHarnessWithFile implements AutoCloseable {
 
 	public StepHarnessWithFile testResource(String filename, String resourceBefore, String resourceAfter) {
 		String contentBefore = ResourceHarness.getTestResource(resourceBefore);
-		File file = harness.setFile(filename).toContent(contentBefore);
-		return test(file, contentBefore, ResourceHarness.getTestResource(resourceAfter));
+		return test(filename, contentBefore, ResourceHarness.getTestResource(resourceAfter));
 	}
 
 	/** Asserts that the given elements in the resources directory are transformed as expected. */
@@ -96,7 +100,7 @@ public class StepHarnessWithFile implements AutoCloseable {
 
 	public AbstractStringAssert<?> testExceptionMsg(File file, String before) {
 		try {
-			formatter.compute(LineEnding.toUnix(before), file);
+			formatter().compute(LineEnding.toUnix(before), file);
 			throw new SecurityException("Expected exception");
 		} catch (Throwable e) {
 			if (e instanceof SecurityException) {
@@ -109,10 +113,5 @@ public class StepHarnessWithFile implements AutoCloseable {
 				return Assertions.assertThat(rootCause.getMessage());
 			}
 		}
-	}
-
-	@Override
-	public void close() {
-		formatter.close();
 	}
 }